# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health": {
            "get": {
                "description": "心跳检测aaa",
                "tags": [
                    "basic"
                ],
                "summary": "心跳检测",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sms": {
            "get": {
                "description": "短信验证码",
                "tags": [
                    "基本模块"
                ],
                "summary": "发送短信验证码",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/changephone": {
            "put": {
                "description": "更换用户手机号, 原手机号验证修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "更换手机号",
                "parameters": [
                    {
                        "description": "更换手机号",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ChangeUserPhoneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/changepwd": {
            "put": {
                "description": "修改用户密码, 通过手机号验证码修改, 通过原密码新密码验证修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ChangeUserPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "获取用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/types.UserInfoResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "用户更新信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserUpdateInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录, 账号密码登录, 手机验证码注册登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录类型, 用户名, 密码, 手机号, 验证码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回包括用户信息,token,过期时间",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/types.UserLoginResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/logoff": {
            "delete": {
                "description": "注销",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户注销",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "delete": {
                "description": "退出登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户退出登录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/recoverpwd": {
            "put": {
                "description": "找回密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "找回密码",
                "parameters": [
                    {
                        "description": "找回密码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RecoverUserPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "system.User": {
            "type": "object",
            "properties": {
                "allVolume": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayAllVolume": {
                    "type": "string"
                },
                "displayUsedVolume": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enableAiPhotoFace": {
                    "type": "boolean"
                },
                "enableAiPhotoLabel": {
                    "type": "boolean"
                },
                "headerImg": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "loginAt": {
                    "type": "integer"
                },
                "nickName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "registerType": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usedVolume": {
                    "type": "integer"
                },
                "usedVolumePercent": {
                    "type": "number"
                },
                "userUID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.ChangeUserPasswordRequest": {
            "type": "object",
            "properties": {
                "changeType": {
                    "type": "string",
                    "example": "修改类型, 原密码修改(password), 手机验证码修改(phone)"
                },
                "newPassword": {
                    "type": "string",
                    "example": "新密码"
                },
                "newPasswordAgain": {
                    "type": "string",
                    "example": "新密码确认"
                },
                "oldPassword": {
                    "type": "string",
                    "example": "旧密码"
                },
                "phoneCode": {
                    "type": "string",
                    "example": "手机号验证码"
                }
            }
        },
        "types.ChangeUserPhoneRequest": {
            "type": "object",
            "properties": {
                "newPhone": {
                    "type": "string",
                    "example": "新手机号码"
                },
                "newPhoneCode": {
                    "type": "string",
                    "example": "新手机验证码"
                },
                "oldPhoneCode": {
                    "type": "string",
                    "example": "旧手机的验证码"
                }
            }
        },
        "types.RecoverUserPasswordRequest": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "type": "string",
                    "example": "新密码"
                },
                "newPasswordAgain": {
                    "type": "string",
                    "example": "新密码确认"
                },
                "phone": {
                    "type": "string",
                    "example": "手机号"
                },
                "phoneCode": {
                    "type": "string",
                    "example": "手机号验证码"
                }
            }
        },
        "types.UserInfoResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/system.User"
                }
            }
        },
        "types.UserLoginRequest": {
            "type": "object",
            "required": [
                "loginType"
            ],
            "properties": {
                "loginType": {
                    "description": "手机验证码或账号密码",
                    "type": "string",
                    "example": "登录类型 必填,phone/account"
                },
                "password": {
                    "description": "账户密码登录, 密码",
                    "type": "string",
                    "example": "密码"
                },
                "phone": {
                    "description": "手机号验证码登录, 手机号码",
                    "type": "string",
                    "example": "手机号"
                },
                "phoneCode": {
                    "description": "手机验证码登录, 验证码",
                    "type": "string",
                    "example": "手机验证码"
                },
                "username": {
                    "description": "账户密码登录, 用户名, 用户名支持手机号密码登录",
                    "type": "string",
                    "example": "用户名"
                }
            }
        },
        "types.UserLoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/system.User"
                }
            }
        },
        "types.UserRegisterRequest": {
            "type": "object",
            "required": [
                "passWord",
                "phone",
                "phoneCode",
                "username"
            ],
            "properties": {
                "headerImg": {
                    "type": "string",
                    "example": "头像链接"
                },
                "nickName": {
                    "type": "string",
                    "example": "昵称"
                },
                "passWord": {
                    "type": "string",
                    "example": "密码"
                },
                "phone": {
                    "type": "string",
                    "example": "电话号码, 必填"
                },
                "phoneCode": {
                    "description": "手机验证码登录, 验证码",
                    "type": "string",
                    "example": "手机号验证码"
                },
                "username": {
                    "type": "string",
                    "example": "用户名"
                }
            }
        },
        "types.UserUpdateInfoRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "四川"
                },
                "country": {
                    "type": "string",
                    "example": "中国"
                },
                "headerImg": {
                    "type": "string",
                    "example": "头像链接"
                },
                "nickName": {
                    "description": "ID        uint   `json:\"-\"`",
                    "type": "string",
                    "example": "这是昵称"
                },
                "province": {
                    "type": "string",
                    "example": "成都"
                },
                "sex": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}